+++ Magic/CHANGELOG.md
@@ -10,6 +10,7 @@
  - Support attribute equations in wand properties and configs
  - Fix incompatibilities with certain GUI plugins such as Jobs Reloaded
  - Support custom items as spell/wand icons (doesn't really do much except let you define them in items.yml)
+ - Fix Capture when used on pigmen
 
 # 7.5.4